+++ lib/components/refrigerator/camera_image_preview.dart
