+++ lib/components/refrigerator/camera_shutter.dart
