+++ lib/components/refrigerator/image_confirm_select.dart
