+++ lib/utils/app_router.dart
@@ -6,12 +6,12 @@ import 'package:nutripic/models/diary_model.dart';
 import 'package:nutripic/models/recipe_model.dart';
 import 'package:nutripic/models/refrigerator_model.dart';
 import 'package:nutripic/models/user_model.dart';
-import 'package:nutripic/view_models/camera/camera_add_view_model.dart';
-import 'package:nutripic/view_models/camera/camera_confirm_view_model.dart';
-import 'package:nutripic/view_models/camera/camera_loading_fail_view_model.dart';
-import 'package:nutripic/view_models/camera/camera_loading_view_model.dart';
-import 'package:nutripic/view_models/camera/camera_view_model.dart';
-import 'package:nutripic/view_models/camera/food_edit_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_add_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_confirm_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_loading_fail_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_loading_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/food_edit_view_model.dart';
 import 'package:nutripic/view_models/diary/diary_post_view_model.dart';
 import 'package:nutripic/view_models/diary/diary_record_view_model.dart';
 import 'package:nutripic/view_models/diary/diary_view_model.dart';
@@ -27,16 +27,16 @@ import 'package:nutripic/view_models/onboarding_view_model.dart';
 import 'package:nutripic/view_models/refrigerator/refrigerator_view_model.dart';
 import 'package:nutripic/view_models/user_info/user_edit_view_model.dart';
 import 'package:nutripic/view_models/user_info/user_info_view_model.dart';
-import 'package:nutripic/views/camera/camera_add_view.dart';
-import 'package:nutripic/views/camera/camera_confirm_view.dart';
-import 'package:nutripic/views/camera/camera_loading_fail_view.dart';
-import 'package:nutripic/views/camera/camera_loading_view.dart';
-import 'package:nutripic/views/camera/food_edit_view.dart';
+import 'package:nutripic/views/refrigerator/camera_add_view.dart';
+import 'package:nutripic/views/refrigerator/camera_confirm_view.dart';
+import 'package:nutripic/views/refrigerator/camera_loading_fail_view.dart';
+import 'package:nutripic/views/refrigerator/camera_loading_view.dart';
+import 'package:nutripic/views/refrigerator/food_edit_view.dart';
 import 'package:nutripic/views/diary/diary_record_view.dart';
 import 'package:nutripic/views/diary/diary_view.dart';
 import 'package:nutripic/views/login/email_view.dart';
 import 'package:nutripic/views/login/signup_view.dart';
-import 'package:nutripic/views/camera/camera_view.dart';
+import 'package:nutripic/views/refrigerator/camera_view.dart';
 import 'package:nutripic/views/diary/diary_post_view.dart';
 import 'package:nutripic/views/login/login_view.dart';
 import 'package:nutripic/views/recipe/recipe_detail_view.dart';
