+++ lib/view_models/refrigerator/camera_add_view_model.dart
