+++ lib/view_models/refrigerator/camera_confirm_view_model.dart
