+++ lib/view_models/refrigerator/camera_loading_fail_view_model.dart
