+++ lib/view_models/refrigerator/camera_loading_view_model.dart
