+++ lib/view_models/refrigerator/camera_view_model.dart
