+++ lib/views/refrigerator/camera_confirm_view.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:nutripic/components/camera/image_confirm_grid.dart';
+import 'package:nutripic/components/refrigerator/image_confirm_grid.dart';
 import 'package:nutripic/components/common/custom_app_bar.dart';
 import 'package:nutripic/components/common/custom_scaffold.dart';
-import 'package:nutripic/components/main_button.dart';
-import 'package:nutripic/components/box_button.dart';
+import 'package:nutripic/components/common/main_button.dart';
+import 'package:nutripic/components/common/box_button.dart';
 import 'package:nutripic/utils/enums/box_button_type.dart';
 import 'package:nutripic/utils/enums/main_button_type.dart';
-import 'package:nutripic/view_models/camera/camera_confirm_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_confirm_view_model.dart';
 import 'package:provider/provider.dart';
 
 class CameraConfirmView extends StatelessWidget {
