+++ lib/views/refrigerator/camera_loading_fail_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:nutripic/components/box_button.dart';
+import 'package:nutripic/components/common/box_button.dart';
 import 'package:nutripic/components/common/custom_scaffold.dart';
 import 'package:nutripic/utils/enums/box_button_type.dart';
 import 'package:nutripic/utils/palette.dart';
-import 'package:nutripic/view_models/camera/camera_loading_fail_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_loading_fail_view_model.dart';
 import 'package:provider/provider.dart';
 
 class CameraLoadingFailView extends StatelessWidget {
