+++ lib/views/refrigerator/camera_loading_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:nutripic/components/common/custom_scaffold.dart';
 import 'package:nutripic/utils/palette.dart';
-import 'package:nutripic/view_models/camera/camera_loading_view_model.dart';
+import 'package:nutripic/view_models/refrigerator/camera_loading_view_model.dart';
 import 'package:provider/provider.dart';
 
 class CameraLoadingView extends StatelessWidget {
